split plugin.py into smaller files (#427)
